refactor(l1): split staffing validation helpers
